refactor: streamline ADN request forms by consolidating state management and API calls into a custom hook
