+++ dna_fe/src/components/ADNRequestFormCivil.jsx
@@ -1,144 +1,24 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { fetchAccountInfo, getCustomerByAccountId } from '../api/accountApi';
+import React from 'react';
 import { getAllCivilServices } from '../api/serviceApi';
+import useADNRequestForm from '../hooks/useADNForm';
 import '../styles/components/ADNRequestForm.css';
 
 const ADNRequestCivilForm = () => {
-  const [customer, setCustomer] = useState({
-    requesterName: '',
-    gender: 'Nam',
-    idCard: '',
-    issueDate: '',
-    issuePlace: '',
-    address: '',
-    phone: '',
-    email: ''
-  });
-
-  const [sampleCount, setSampleCount] = useState("");
-  const [isLoading, setIsLoading] = useState(true);
-  const [services, setServices] = useState([]);
-  const [selectedService, setSelectedService] = useState(null);
-  const [formData, setFormData] = useState({});
-  const [error, setError] = useState(null);
-  const [success, setSuccess] = useState(null);
-
-  const formatDate = (iso) => iso?.split('T')[0] || '';
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const accRes = await fetchAccountInfo();
-        const acc = accRes.data;
-        const cusRes = await getCustomerByAccountId(acc.id);
-        const cus = cusRes.data;
-
-        setCustomer({
-          requesterName: acc.fullName || '',
-          gender: cus.gender || 'Nam',
-          idCard: cus.cccd || '',
-          issueDate: formatDate(cus.dateOfIssue),
-          issuePlace: cus.placeOfIssue || '',
-          address: cus.address || '',
-          phone: acc.phone || '',
-          email: acc.email || ''
-        });
-
-        const civilServices = await getAllCivilServices();
-        setServices(civilServices);
-      } catch (err) {
-        console.error('Error fetching info:', err);
-        setError('Failed to load customer or service data.');
-      } finally {
-        setIsLoading(false);
-      }
-    };
-    fetchData();
-  }, []);
-
-  const handleSampleChange = (e) => {
-    const value = e.target.value;
-    setSampleCount(value ? parseInt(value) : "");
-  };
-
-  const handleInputChange = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setError(null);
-    setSuccess(null);
-
-    try {
-      const customerId = localStorage.getItem('customerId');
-      if (!customerId) {
-        setError('Customer ID not found in localStorage. Please log in.');
-        return;
-      }
-
-      const orderData = {
-        customerId: parseInt(customerId),
-        staffId: null,
-        serviceId: parseInt(formData.testType),
-        orderDate: formData.orderDate || new Date().toISOString().split('T')[0],
-        orderStatus: formData.orderStatus || 'PENDING',
-        sampleType: formData.method || '',
-        resultDeliveryMethod: formData.receiveAt || '',
-        resultDeliverAddress: formData.resultAddress || '',
-        kitCode: formData.method === 'home' ? `KIT-${Math.random().toString(36).substr(2, 9)}` : null,
-        sampleQuantity: parseInt(sampleCount) || 2,
-        amount: calculateTotalPrice() || 0,
-      };
-
-      const orderResponse = await axios.post('http://localhost:8080/api/testorders', orderData, {
-        headers: { 'Content-Type': 'application/json' },
-      });
-      const orderId = orderResponse.data.orderId;
-
-      const samples = [];
-      for (let i = 1; i <= sampleCount; i++) {
-        const index = i;
-        const sampleData = {
-          orderId,
-          customerId: parseInt(customerId),
-          staffId: null,
-          name: formData[`person${index}Name`] || '',
-          gender: formData[`person${index}Gender`] || '',
-          dateOfBirth: formData[`person${index}Dob`] || '',
-          relationship: formData[`person${index}Relationship`] || '',
-          sampleType: formData[`person${index}SampleType`] || '',
-          numberOfSample: 1,
-          medicalHistory: formData[`person${index}BloodHistory`] || '',
-          fingerprint: '',
-        };
-        samples.push(sampleData);
-      }
-
-      const failedSamples = [];
-      for (const [index, sample] of samples.entries()) {
-        try {
-          await axios.post('http://localhost:8080/api/testSamples', sample, {
-            headers: { 'Content-Type': 'application/json' },
-          });
-        } catch (sampleErr) {
-          failedSamples.push(`Sample ${index + 1}: ${sampleErr.response?.data?.message || sampleErr.message}`);
-        }
-      }
-
-      if (failedSamples.length > 0) {
-        setError(`Some samples failed to save: ${failedSamples.join('; ')}`);
-      } else {
-        setSuccess('Bạn đã tạo đơn thành công');
-        setFormData({});
-        setSampleCount("");
-      }
-    } catch (err) {
-      setError(err.response?.data?.message || 'Failed to create Test Order or Samples.');
-      console.error('Error:', err);
-    }
-  };
+  const {
+    customer,
+    sampleCount,
+    isLoading,
+    services,
+    selectedService,
+    setSelectedService,
+    formData,
+    error,
+    success,
+    handleSampleChange,
+    handleInputChange,
+    handleSubmit,
+    calculateTotalPrice,
+  } = useADNRequestForm(getAllCivilServices);
 
   const renderSampleFields = () => {
     const fields = [];
@@ -220,13 +100,7 @@ const ADNRequestCivilForm = () => {
     return fields;
   };
 
-  const calculateTotalPrice = () => {
-    if (!selectedService || !sampleCount || isNaN(sampleCount)) return null;
-    const additionalCost = (sampleCount - 2) * 1500000;
-    return selectedService.price + (additionalCost > 0 ? additionalCost : 0);
-  };
-
-  const totalPrice = calculateTotalPrice();
+  const totalPrice = calculateTotalPrice(true);
 
   if (isLoading) return <div className="adn-loading">Đang tải thông tin...</div>;
 
@@ -235,7 +109,7 @@ const ADNRequestCivilForm = () => {
       <h2 className="adn-form-title">ĐƠN YÊU CẦU PHÂN TÍCH ADN DÂN SỰ</h2>
       {error && <div className="text-red-500 mb-4">{error}</div>}
       {success && <div className="text-green-500 mb-4">{success}</div>}
-      <form className="adn-form" onSubmit={handleSubmit}>
+      <form className="adn-form" onSubmit={(e) => handleSubmit(e, true)}>
         <label>Họ tên người yêu cầu:</label>
         <input type="text" name="requesterName" value={customer.requesterName} disabled />
